requirements.txt
asgiref==3.3.2
cffi==1.14.5
cryptography==3.4.7
Django==3.1.7
django-filter==2.4.0
django-rest-knox==4.1.0
django-rest-passwordreset==1.1.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
importlib-metadata==3.10.0
Markdown==3.3.4
mysqlclient==2.0.3
Pillow==8.2.0
pycparser==2.20
PyJWT==1.7.1
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
zipp==3.4.1